Fix for event tracking URL overwriting v1 API URL.
This issue was evident when running an event tracking API request, then running a version 1 API request. The error for the v1 request would be: 'An account ID must be provided.' It was still holding on to the tracking URL as the main URL.

This change stops overwriting the main URL and instead uses the tracking URL only when needed.

DEVJIRA-8782